feat: Add SQLite database support and Admin Abuse Logging system
