+++ src/main/java/org/ssoggy/ssoggysouls/command/ReviveCommand.java
@@ -18,6 +18,7 @@ import org.ssoggy.ssoggysouls.util.MessageUtil;
 import org.ssoggy.ssoggysouls.util.PermissionUtil;
 import org.ssoggy.ssoggysouls.util.PlayerRevivalUtil;
 import org.ssoggy.ssoggysouls.util.TabCompleteUtil;
+import org.ssoggy.ssoggysouls.util.AdminLogger;
 
 public class ReviveCommand implements CommandExecutor, TabCompleter {
 
@@ -69,8 +70,7 @@ public class ReviveCommand implements CommandExecutor, TabCompleter {
             boolean success = db.revivePlayer(data.getUuid(), livesToRestore);
 
             if (success) {
-                plugin.getLogger().log(Level.INFO, "{0} revived {1} (lives: {2})",
-                        new Object[]{sender.getName(), data.getUsername(), livesToRestore});
+                AdminLogger.log(plugin, sender.getName(), "revived " + data.getUsername() + " (lives: " + livesToRestore + ")");
                 sender.sendMessage(MessageUtil.get("revive-admin-success",
                         KEY_PLAYER, data.getUsername(),
                         "lives", livesToRestore));
